+++ lib/load.php
@@ -226,3 +226,8 @@ if ( gutenberg_is_experiment_enabled( 'gutenberg-content-types' ) ) {
 	require __DIR__ . '/experimental/content-types/load.php';
 	require __DIR__ . '/experimental/content-types/index.php';
 }
+
+// Dashboard Widgets (only load when experiment is enabled).
+if ( gutenberg_is_experiment_enabled( 'gutenberg-dashboard-widgets' ) ) {
+	require __DIR__ . '/experimental/dashboard-widgets/load.php';
+}