+++ packages/editor/src/components/post-card-panel/index.js
@@ -6,7 +6,7 @@ import {
 	Button,
 	__experimentalHStack as HStack,
 	__experimentalVStack as VStack,
-	__experimentalText as Text,
+	__experimentalText as WCText,
 	privateApis as componentsPrivateApis,
 } from '@wordpress/components';
 import { close } from '@wordpress/icons';
@@ -28,7 +28,7 @@ import { unlock } from '../../lock-unlock';
 import PostActions from '../post-actions';
 import usePageTypeBadge from '../../utils/pageTypeBadge';
 import { getTemplateInfo } from '../../utils/get-template-info';
-const { Badge } = unlock( componentsPrivateApis );
+const { Badge: WCBadge } = unlock( componentsPrivateApis );
 
 /**
  * Renders a title of the post type and the available quick actions available within a 3-dot dropdown.
@@ -131,7 +131,7 @@ export default function PostCardPanel( {
 				alignment="flex-start"
 			>
 				<Icon className="editor-post-card-panel__icon" icon={ icon } />
-				<Text
+				<WCText
 					numberOfLines={ 2 }
 					truncate
 					className="editor-post-card-panel__title"
@@ -141,9 +141,9 @@ export default function PostCardPanel( {
 						{ title }
 					</span>
 					{ pageTypeBadge && postIds.length === 1 && (
-						<Badge>{ pageTypeBadge }</Badge>
+						<WCBadge>{ pageTypeBadge }</WCBadge>
 					) }
-				</Text>
+				</WCText>
 				{ ! hideActions && postIds.length === 1 && (
 					<PostActions
 						postType={ postType }
@@ -161,13 +161,13 @@ export default function PostCardPanel( {
 				) }
 			</HStack>
 			{ postIds.length > 1 && (
-				<Text className="editor-post-card-panel__description">
+				<WCText className="editor-post-card-panel__description">
 					{ sprintf(
 						// translators: %s: Name of the plural post type e.g: "Posts".
 						__( 'Changes will be applied to all selected %s.' ),
 						labels?.name.toLowerCase()
 					) }
-				</Text>
+				</WCText>
 			) }
 		</VStack>
 	);