+++ routes/dashboard/stage.tsx
@@ -1,0 +1,15 @@
+/**
+ * WordPress dependencies
+ */
+import { Page } from '@wordpress/admin-ui';
+import { __ } from '@wordpress/i18n';
+
+function Dashboard() {
+	return (
+		<Page title={ __( 'Dashboard' ) } headingLevel={ 1 }>
+			<div className="dashboard-widgets" />
+		</Page>
+	);
+}
+
+export const stage = Dashboard;