+++ routes/taxonomies/stage.tsx
@@ -11,7 +11,11 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import AddTaxonomy from './add-taxonomy';
-import { editTaxonomyAction, toggleActiveAction } from './actions';
+import {
+	deleteTaxonomyAction,
+	editTaxonomyAction,
+	toggleActiveAction,
+} from './actions';
 import {
 	titleField,
 	statusField,
@@ -20,7 +24,9 @@ import {
 	useSlugField,
 	useObjectTypeField,
 } from './fields';
-import { toFormData, type TaxonomyRecord } from './utils';
+import { toFormData } from './utils';
+import type { TaxonomyRecord } from './types';
+import './style.scss';
 
 const defaultLayouts = {
 	table: {},
@@ -38,7 +44,7 @@ const DEFAULT_VIEW: View = {
 function TaxonomiesPage() {
 	const [ view, setView ] = useState< View >( DEFAULT_VIEW );
 	const taxonomyActions = useMemo(
-		() => [ editTaxonomyAction, toggleActiveAction ],
+		() => [ editTaxonomyAction, toggleActiveAction, deleteTaxonomyAction ],
 		[]
 	);
 	const slugField = useSlugField();
@@ -86,7 +92,12 @@ function TaxonomiesPage() {
 		[ totalItems, totalPages ]
 	);
 	return (
-		<Page title={ __( 'Taxonomies' ) } actions={ <AddTaxonomy /> }>
+		<Page
+			title={ __( 'Taxonomies' ) }
+			className="taxonomies-page"
+			hasPadding={ false }
+			actions={ <AddTaxonomy /> }
+		>
 			<DataViews
 				data={ data }
 				fields={ fields }